move websocket server logic into its own module
