+++ src/backend/app/__init__.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from socketio import ASGIApp
 from app.routing import api_router
-from app.websocket import socket
+from app.websocket import socket_server
 from starlette.middleware.cors import CORSMiddleware
 
 origins = [
@@ -22,7 +22,7 @@ app.add_middleware(
 app.include_router(api_router, prefix="/api")
 
 socket_app = ASGIApp(
-    socketio_server=socket, 
+    socketio_server=socket_server, 
     other_asgi_app=app,
     socketio_path="socket.io"
 )
